+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Added:
 Fixed:
 ^^^^^^
 - HiLo look up table max is now 4095
+- Use of platemap SVGs with green well active indicators
 
 
 0.1.36 (2023-07-31)
@@ -27,8 +28,6 @@ Added:
 - Async stage dialog control do ui isn't blocked when moving stage
 
 
-Added:
-^^^^^^
 0.1.35 (2023-07-27)
 -------------------